Add external archive loading API for bundle storage view
Move external archive loading logic no NestBuildRepositoryImpl.
