+++ src/saker/nest/bundle/AbstractExternalArchive.java
@@ -20,5 +20,5 @@ import java.io.Closeable;
 public abstract class AbstractExternalArchive implements ExternalArchive, Closeable {
 	public abstract byte[] getSharedHash();
 
-	public abstract ExternalArchiveKey getArchiveKey();
+	public abstract SimpleExternalArchiveKey getArchiveKey();
 }
